% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{bleu}
\alias{bleu}
\title{BLEU  (Bilingual Evaluation Understudy)}
\usage{
bleu(.data, n = 4)
}
\arguments{
\item{.data}{A list, where each candidate has multiple references  (dataframe that was transformed into a list by another function in this package)}

\item{n}{desired number of ngrams (defaults to 4)}
}
\value{
tibble with data in long form. Each row contains the BLEU score for each ngram value for each candidate-reference pair.
}
\description{
BLEU score calculations over a data frame.
}
\examples{
# Our candidate and reference data
sample_data <- list(list(candidate = "The cat the cat on the mat", references = c("The cat is on the mat", "There is a cat on the mat")),
list(candidate = "The cat the cat on the map", references = c("The cat is on the ccccccat", "There is a cat on the mat")))



# Run function
bleu(sample_data)
}
