% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brev_pen.R
\name{brev_pen}
\alias{brev_pen}
\title{BLEU  (Bilingual Evaluation Understudy)}
\usage{
brev_pen(data, reference = reference, candidate = candidate)
}
\arguments{
\item{data}{data frame (data that was processed by a data loading function to be created)}

\item{reference}{column in data with reference text (human translated or created text)}

\item{candidate}{column in data with candidate text (machine generated text)}
}
\value{
brevity penalty score calculated as e^(1-r/c)
}
\description{
Brevity Penalty calculations for BLEU assessed candidate and reference texts
}
\examples{
library(footrulr)

brev_pen(data = wmt18example, reference = "reference", candidate = "candidate")
}
