% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footrulr-package.R
\docType{package}
\name{footrulr-package}
\alias{footrulr}
\alias{footrulr-package}
\title{footrulr: Measure Quality of Machine Generated Texts}
\description{
The goal of footrulr is to provide some
    functionality to compute sentence comparisons using metrics frequently
    employed in Machine Translation and Text Summarization evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kanishkamisra/footrulr}
  \item Report bugs at \url{https://github.com/kanishkamisra/footrulr/issues}
}

}
\author{
\strong{Maintainer}: Kanishka Misra \email{kmisra@purdue.edu}

Other contributors:
\itemize{
  \item William Bonnell \email{wtb3na@virginia.edu} [contributor]
  \item Joshua Goldberg \email{josh.goldberg1@outlook.com} [contributor]
  \item Mauro Lepore \email{maurolepore@gmail.com} (0000-0002-1986-7988) [contributor]
  \item Katherine Simeon \email{katherine@rladies.org} [contributor]
}

}
\keyword{internal}
