% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footrulr.R
\name{footrulr}
\alias{footrulr}
\title{footrulr}
\usage{
footrulr(cand_file, ..., non_value = "NO REFERENCE AVAILABLE")
}
\arguments{
\item{cand_file}{candidate file}

\item{...}{one or more reference files}

\item{non_value}{string indicating non references}
}
\value{
a footrulr object with candidate-reference collections
}
\description{
footrulr
}
