% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_text.R
\name{load_text}
\alias{load_text}
\title{load_text}
\usage{
load_text(candidate_df, reference_df)
}
\arguments{
\item{candidate_df}{candidate text file}

\item{reference_df}{reference text file}
}
\value{
a dataframe with candidates matched to references
}
\description{
Load two text files (a candidate and a reference) and give it unique candidate and reference IDs and pair them.
Candidates and references should have a one-to-one correspondence
}
