% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wmt18example}
\alias{wmt18example}
\title{wmt18example}
\format{A data frame with 24288 rows and 5 variables:
\describe{
\item{\code{system}}{character COLUMN_DESCRIPTION}
\item{\code{candidate_id}}{integer COLUMN_DESCRIPTION}
\item{\code{reference_id}}{integer COLUMN_DESCRIPTION}
\item{\code{candidate}}{character COLUMN_DESCRIPTION}
\item{\code{reference}}{character COLUMN_DESCRIPTION}
}}
\usage{
wmt18example
}
\description{
Example references
}
\details{
DETAILS
}
\keyword{datasets}
