% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{bleu}
\alias{bleu}
\title{BLEU  (Bilingual Evaluation Understudy)}
\usage{
bleu(.data, n = 4)
}
\arguments{
\item{.data}{data frame (data that was processed by a data loading function to be created)}

\item{n}{desired number of ngrams (defaults to 4)}
}
\value{
tibble with data in long form. Each row contains the BLEU score for each ngram value for each candidate-reference pair.
}
\description{
BLEU score calculations over a data frame.
}
\examples{
# Our candidate and reference data
candidate <- "The cat the cat on the mat"
reference1 <- "The cat is on the mat"
reference2 <- "There is a cat on the mat"

# Create data frame
pairs <- data.frame(cbind(
  candidate = c(candidate, candidate), reference = c(reference1, reference2)
))
sample_data <- cbind(system = c(1, 1), candidate_id = c(1, 1), reference_id = c(1, 2), pairs)

# Run function
bleu(sample_data)
}
