% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footrulr-package.R
\docType{package}
\name{footrulr-package}
\alias{footrulr}
\alias{footrulr-package}
\title{footrulr: Compare sentences using Machine Translation and Text Summarization evaluation metrics}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kanishkamisra/footrulr}
  \item Report bugs at \url{https://github.com/kanishkamisra/footrulr/issues}
}

}
\author{
\strong{Maintainer}: Kanishka Misra \email{kmisra@purdue.edu}

}
\keyword{internal}
