% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wmt18example}
\alias{wmt18example}
\title{wmt18example}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24288 rows and 4 columns.}
\usage{
wmt18example
}
\description{
wmt18example
}
\keyword{datasets}
